+++ src/Concerns/Constructor.php
@@ -6,7 +6,7 @@ use Henzeb\Enumhancer\Helpers\EnumMakers;
 
 trait Constructor
 {
-    final public static function __callStatic(string $name, array $arguments)
+    public static function __callStatic(string $name, array $arguments)
     {
         return EnumMakers::make(self::class, $name, true);
     }