+++ src/Concerns/Extractor.php
@@ -7,7 +7,7 @@ use Henzeb\Enumhancer\Helpers\EnumExtractor;
 
 trait Extractor
 {
-    final public static function extract(string $text, Mapper|string|null $mapper = null): array
+    public static function extract(string $text, Mapper|string|null $mapper = null): array
     {
         return EnumExtractor::extract(self::class, $text, $mapper);
     }