+++ src/Concerns/Makers.php
@@ -6,12 +6,12 @@ use Henzeb\Enumhancer\Helpers\EnumMakers;
 
 trait Makers
 {
-    final public static function make(int|string|null $value): self
+    public static function make(int|string|null $value): self
     {
         return EnumMakers::make(self::class, $value);
     }
 
-    final public static function tryMake(int|string|null $value): ?self
+    public static function tryMake(int|string|null $value): ?self
     {
         return EnumMakers::tryMake(self::class, $value);
     }
@@ -20,7 +20,7 @@ trait Makers
      * @param iterable $values
      * @return self[]
      */
-    final public static function makeArray(iterable $values): array
+    public static function makeArray(iterable $values): array
     {
         return EnumMakers::makeArray(self::class, $values);
     }
@@ -29,7 +29,7 @@ trait Makers
      * @param iterable $values
      * @return self[]
      */
-    final public static function tryMakeArray(iterable $values): array
+    public static function tryMakeArray(iterable $values): array
     {
         return EnumMakers::tryMakeArray(self::class, $values);
     }