+++ src/Concerns/Subset.php
@@ -7,7 +7,7 @@ use Henzeb\Enumhancer\Helpers\EnumSubsetMethods;
 
 trait Subset
 {
-    final public static function of(self ...$enums): EnumSubset
+    public static function of(self ...$enums): EnumSubset
     {
         return new EnumSubsetMethods(self::class, ...($enums ?:self::cases()));
     }