+++ tests/Unit/Concerns/ValueTest.php
@@ -5,6 +5,7 @@ namespace Henzeb\Enumhancer\Tests\Unit\Concerns;
 use PHPUnit\Framework\TestCase;
 use Henzeb\Enumhancer\Tests\Fixtures\EnhancedUnitEnum;
 use Henzeb\Enumhancer\Tests\Fixtures\EnhancedBackedEnum;
+use Henzeb\Enumhancer\Tests\Fixtures\BackedEnums\Value\ValueIntBackedEnum;
 
 
 class ValueTest extends TestCase
@@ -16,6 +17,18 @@ class ValueTest extends TestCase
             'backed-2' => [EnhancedBackedEnum::ANOTHER_ENUM, EnhancedBackedEnum::ANOTHER_ENUM->value],
             'unit-1' => [EnhancedUnitEnum::ENUM, strtolower(EnhancedUnitEnum::ENUM->name)],
             'unit-2' => [EnhancedUnitEnum::ANOTHER_ENUM, strtolower(EnhancedUnitEnum::ANOTHER_ENUM->name)],
+        ];
+    }
+
+    public function providesEnumsForKey()
+    {
+        return [
+            'string-backed-1' => [EnhancedBackedEnum::ENUM, 0],
+            'string-backed-2' => [EnhancedBackedEnum::ANOTHER_ENUM, 1],
+            'int-backed-1' => [ValueIntBackedEnum::ENUM, 64],
+            'int-backed-2' => [ValueIntBackedEnum::ANOTHER_ENUM, 128],
+            'unit-1' => [EnhancedUnitEnum::ENUM, 0],
+            'unit-2' => [EnhancedUnitEnum::ANOTHER_ENUM, 1],
         ];
     }
 
@@ -32,6 +45,21 @@ class ValueTest extends TestCase
         $this->assertEquals(
             $expected,
             $enum->value()
+        );
+    }
+
+    /**
+     * @param $enum
+     * @param string $expected
+     * @return void
+     *
+     * @dataProvider providesEnumsForKey
+     */
+    public function testEnumShouldReturnKey($enum, int $expected): void
+    {
+        $this->assertEquals(
+            $expected,
+            $enum->key()
         );
     }
 }