+++ tests/Unit/Helpers/EnumReporterTest.php
@@ -2,27 +2,35 @@
 
 namespace Henzeb\Enumhancer\Tests\Unit\Helpers;
 
+use Mockery;
+use stdClass;
 use BackedEnum;
+use RuntimeException;
+use Psr\Log\LoggerInterface;
+use Orchestra\Testbench\TestCase;
+use Illuminate\Support\Facades\Log;
+use Henzeb\Enumhancer\Enums\LogLevel;
 use Henzeb\Enumhancer\Contracts\Reporter;
 use Henzeb\Enumhancer\Helpers\EnumReporter;
 use Henzeb\Enumhancer\Laravel\Reporters\LaravelLogReporter;
 
-use PHPUnit\Framework\TestCase;
-use RuntimeException;
-use stdClass;
 
 class EnumReporterTest extends TestCase
 {
-    public function testReporterNotSet() {
+    public function testReporterNotSet()
+    {
         EnumReporter::set(null);
         $this->assertNull(EnumReporter::get());
     }
 
-    public function testSetReporter() {
+    public function testSetReporter()
+    {
         EnumReporter::set(null);
 
         $reporter = new class implements Reporter {
-            public function report(string $enum, ?string $key, ?BackedEnum $context): void {}
+            public function report(string $enum, ?string $key, ?BackedEnum $context): void
+            {
+            }
         };
 
         EnumReporter::set($reporter);
@@ -30,28 +38,32 @@ class EnumReporterTest extends TestCase
         $this->assertEquals($reporter, EnumReporter::get());
     }
 
-    public function testSetStringReporter() {
+    public function testSetStringReporter()
+    {
         EnumReporter::set(null);
         EnumReporter::set(LaravelLogReporter::class);
 
         $this->assertEquals(new LaravelLogReporter(), EnumReporter::get());
     }
 
-    public function testSetNull() {
+    public function testSetNull()
+    {
 
         EnumReporter::set(null);
 
         $this->assertEquals(null, EnumReporter::get());
     }
 
-    public function testIsNotAReporter() {
+    public function testIsNotAReporter()
+    {
 
         $this->expectException(RuntimeException::class);
 
         EnumReporter::set(stdClass::class);
     }
 
-    public function testObjectIsNotAReporter() {
+    public function testObjectIsNotAReporter()
+    {
 
         $this->expectError();
 
@@ -66,12 +78,66 @@ class EnumReporterTest extends TestCase
         $this->assertEquals(LaravelLogReporter::class, EnumReporter::get()::class);
     }
 
-    public function testMakeOrReportShouldErrorWithNonEnum() {
+    public function testSetLaravelReporterWithDifferentLogLevel()
+    {
+        EnumReporter::set(null);
+        EnumReporter::laravel(LogLevel::Alert);
+
+        $spy = Mockery::spy(LoggerInterface::class);
+
+        Log::shouldReceive('stack')
+            ->once()
+            ->with(['stack'])
+            ->andReturn($spy);
+
+        EnumReporter::get()->report(LogLevel::class, null, null);
+
+        $spy->shouldHaveReceived(
+            'log',
+            [
+                'alert',
+                'LogLevel: A null value was passed',
+                [
+                    'class' => LogLevel::class,
+                ]
+            ]
+        );
+    }
+
+    public function testSetLaravelReporterWithDifferenChannels()
+    {
+        EnumReporter::set(null);
+        EnumReporter::laravel(null, 'bugsnag', 'daily');
+
+        $spy = Mockery::spy(LoggerInterface::class);
+
+        Log::shouldReceive('stack')
+            ->once()
+            ->with(['bugsnag', 'daily'])
+            ->andReturn($spy);
+
+        EnumReporter::get()->report(LogLevel::class, null, null);
+
+        $spy->shouldHaveReceived(
+            'log',
+            [
+                'notice',
+                'LogLevel: A null value was passed',
+                [
+                    'class' => LogLevel::class,
+                ]
+            ]
+        );
+    }
+
+    public function testMakeOrReportShouldErrorWithNonEnum()
+    {
         $this->expectError();
         EnumReporter::makeOrReport(stdClass::class, '', null, new LaravelLogReporter());
     }
 
-    public function testMakeOrReportArrayShouldErrorWithNonEnum() {
+    public function testMakeOrReportArrayShouldErrorWithNonEnum()
+    {
         $this->expectError();
         EnumReporter::makeOrReportArray(stdClass::class, [], null, new LaravelLogReporter());
     }