+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 All notable changes to `Enumhancer` will be documented in this file
 
-## 1.12.0 - 2022-06-19
+## 1.15.0 - 2022-06-21
+
+- Made the Laravel [Reporter](docs/reporters.md#laravel) configurable
+- added `key` method to [Value](docs/value.md)
+
+## 1.14.0 - 2022-06-19
 
 - Added transition hooks [State](docs/state.md)
-- [Makers](docs/makers.md) & [From](docs/from.md) now allow you to use integer 
+- [Makers](docs/makers.md) & [From](docs/from.md) now allow you to use integer
   keys on basic and string enums
 
 ## 1.12.0 - 2022-06-15